+++ tasks/serve/_reload.js
@@ -1,0 +1,14 @@
+import gulp from 'gulp';
+import { server } from '@tasks/server';
+
+/**
+ * Task: serve:reload
+ */
+export function reloadServer(done) {
+  server.reload();
+  done();
+}
+reloadServer.displayName = 'serve:reload';
+reloadServer.description = 'Tell browser to reload the page.';
+
+gulp.task('serve:reload', reloadServer);