+++ tasks/watch/_app.js
@@ -1,8 +1,20 @@
-const gulp = require('gulp');
+import gulp from 'gulp';
+import * as config from '@tasks/config';
+import { buildApp } from '@tasks/build/_app';
+import { reloadServer } from '@tasks/serve/_reload';
 
-function watchApp() {
-  gulp.watch('src/app/**/*.js', ['build:app']);
+function rebuildOnChange() {
+  gulp.watch(`${config.paths.src}/app/**/*.js`, gulp.series(buildApp, reloadServer));
 }
-watchApp.description = 'Rebuild application scripts.';
 
-gulp.task('watch:app', ['build:app'], watchApp);
+/**
+ * Task: watch:app
+ */
+export const watchApp = gulp.series(
+  buildApp,
+  rebuildOnChange
+);
+watchApp.displayName = 'watch:app';
+watchApp.description = 'Rebuild application scripts on change.';
+
+gulp.task(watchApp);