+++ tasks/watch/_styles.js
@@ -1,8 +1,20 @@
-const gulp = require('gulp');
+import gulp from 'gulp';
+import * as config from '@tasks/config';
+import { buildStyles } from '@tasks/build/_styles';
+import { reloadServer } from '@tasks/serve/_reload';
 
-function watchStyles() {
-  gulp.watch('src/styles/**/*.scss', ['build:styles']);
+function rebuildOnChange() {
+  gulp.watch(`${config.paths.src}/styles/**/*.+(sass|scss)`, gulp.series(buildStyles, reloadServer));
 }
-watchStyles.description = 'Recompile Sass files on changes.';
 
-gulp.task('watch:styles', ['build:styles'], watchStyles);
+/**
+ * Task: watch:styles
+ */
+export const watchStyles = gulp.series(
+  buildStyles,
+  rebuildOnChange
+);
+watchStyles.displayName = 'watch:styles';
+watchStyles.description = 'Recompile Sass files on change.';
+
+gulp.task(watchStyles);