+++ tasks/watch/watch.js
@@ -1,15 +1,19 @@
-const gulp = require('gulp');
+import gulp from 'gulp';
+import { watchApp } from '@tasks/watch/_app';
+import { watchAssets } from '@tasks/watch/_assets';
+import { watchStyles } from '@tasks/watch/_styles';
+import { watchViews } from '@tasks/watch/_views';
 
-function watch() {
-  // empty
-}
-watch.description = 'Rebuild application components on file changes.';
+/**
+ * Task: watch
+ */
+export const watch = gulp.parallel(
+  watchApp,
+  watchAssets,
+  watchStyles,
+  watchViews
+);
+watch.displayName = 'watch';
+watch.description = 'Rebuild all application components on change.';
 
-const tasksToRun = [
-  'watch:app',
-  'watch:assets',
-  'watch:styles',
-  'watch:views'
-];
-
-gulp.task('watch', tasksToRun, watch);
+gulp.task(watch);